+++ src/Network/Network/NetworkManager/GetAzNetworkManagerEffectiveVirtualNetworksByNetworkGroupListCommand.cs
@@ -1,0 +1,80 @@
+﻿// ----------------------------------------------------------------------------------
+//
+// Copyright Microsoft Corporation
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ----------------------------------------------------------------------------------
+
+using AutoMapper;
+using Microsoft.Azure.Commands.Network.Models;
+using Microsoft.Azure.Commands.Network.Models.NetworkManager;
+using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
+using Microsoft.Azure.Commands.ResourceManager.Common.Tags;
+using Microsoft.Azure.Management.Network;
+using Microsoft.Azure.Management.Network.Models;
+using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
+using System.Management.Automation;
+using MNM = Microsoft.Azure.Management.Network.Models;
+
+namespace Microsoft.Azure.Commands.Network
+{
+    [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "NetworkManagerEffectiveVirtualNetworkByNetworkGroupList"), OutputType(typeof(PSNetworkManagerEffectiveVirtualNetworksListResult))]
+    public class GetAzNetworkManagerEffectiveVirtualNetworksByNetworkGroupListCommand : NetworkManagerBaseCmdlet
+    {
+        [Parameter(
+           Mandatory = true,
+           ValueFromPipelineByPropertyName = true,
+           HelpMessage = "The networkManager networkgroup name.")]
+        [ValidateNotNullOrEmpty]
+        [ResourceNameCompleter("Microsoft.Network/networkManagers/networkGroups", "ResourceGroupName", "NetworkManagerName")]
+        [SupportsWildcards]
+        public virtual string NetworkGroupName { get; set; }
+
+        [Parameter(
+           Mandatory = true,
+           ValueFromPipelineByPropertyName = true,
+           HelpMessage = "The networkManager name.")]
+        [ValidateNotNullOrEmpty]
+        [ResourceNameCompleter("Microsoft.Network/networkManagers", "ResourceGroupName")]
+        [SupportsWildcards]
+        public virtual string NetworkManagerName { get; set; }
+
+        [Parameter(
+           Mandatory = true,
+           ValueFromPipelineByPropertyName = true,
+           HelpMessage = "The resource group name.")]
+        [ResourceGroupCompleter]
+        [ValidateNotNullOrEmpty]
+        [SupportsWildcards]
+        public virtual string ResourceGroupName { get; set; }
+
+        [Parameter(
+          Mandatory = false,
+          ValueFromPipelineByPropertyName = true,
+          HelpMessage = "SkipToken.")]
+        public string SkipToken { get; set; }
+
+        public override void Execute()
+        {
+            base.Execute();
+            var parameter = new QueryRequestOptions();
+            if (!string.IsNullOrEmpty(this.SkipToken))
+            {
+                parameter.SkipToken = this.SkipToken;
+            }
+
+            var effectiveVirtualNetworksListResult = this.NetworkClient.NetworkManagementClient.ListEffectiveVirtualNetworks.ByNetworkGroup(this.ResourceGroupName, this.NetworkManagerName, this.NetworkGroupName, parameter);
+            var psEffectiveVirtualNetworksList = NetworkResourceManagerProfile.Mapper.Map<PSNetworkManagerEffectiveVirtualNetworksListResult>(effectiveVirtualNetworksListResult);
+            WriteObject(psEffectiveVirtualNetworksList);
+        }
+    }
+}